Add Charts Widget for statistics charts
Checks for presense of Qt5Charts library and Adds a chart widget
that displays various statistic charts in a modal window if the
library is installed.

Adds "Statistics charts" menu option in the "View" menu with a
shortcut of <Alt>-G if the Qt5Charts library is installed

Years and Months bar charts are created from the DateStatsTableModel
data added previously.

Modifies qt59 travis environment to add Qt5Charts library

Signed-off-by: Doug Junkins <junkins@foghead.com>